Separate Robot and DNA (SOC)
