+++ sim_game/robot.py
@@ -14,6 +14,7 @@ except:
 
 from graphics import GraphicsRectangle
 from physics import PhysicsRectangle
+from dna import DNA
 from collections import UserDict
 import random
 import math
@@ -21,112 +22,11 @@ import math
 def distance(x1,y1,x2,y2):
     return math.sqrt((x2-x1)**2 + (y2-y1)**2)
 
-class RobotDNA(UserDict):
-    def __init__(self, randomize=True, **kwargs):
-        self.binary = {  # Binary, true or false genes
-
-        }
-        self.real = {    # Real valued genes
-            "red": 0.5,
-            "green": 0.5,
-            "blue": 0.5,
-            "size": 0.5,
-            "ratio": 0.5
-        }
-        if randomize:
-            self.randomize()
-        self.set(**kwargs)
-
-        self.master = {"binary": self.binary, "real": self.real}
-        super().__init__(self.master)
-
-    def combine(self, dna, probability_average=0.3, weighted_average=True):
-        parents = [self, dna]
-        child = RobotDNA(randomize=False)
-        for key in dna.binary:
-            r = random.randint(0,1)
-            child.binary[key] = parents[r].binary[key]
-
-        for key in dna.real:
-            p = random.uniform(0.0,1.0)
-            if p < probability_average:
-                w1 = 0.5
-                if weighted_average:
-                    w1 = random.uniform(0.0,1.0)
-                w2 = 1.0 - w1
-                a,b = parents[0].real[key], parents[1].real[key]
-                child.real[key] = a * w1 + b * w2
-            else:
-                r = random.randint(0,1)
-                child.real[key] = parents[r].real[key]
-        return child
-
-
-    def set(self, **kwargs):
-        for key, value in kwargs.items():
-            if type(value) is bool:
-                if key in self.binary:
-                    self.binary[key] = value
-                else:
-                    raise KeyError
-            else:
-                if key in self.real:
-                    assert value >= 0 and value <= 1
-                    self.real[key] = float(value)
-                else:
-                    raise KeyError
-
-    def get(self, key):
-        # Returns a value or raises KeyError:
-        if key not in self.binary:
-            return self.real[key]
-        if key not in self.real:
-            return self.binary[key]
-
-        # key in both dicts:
-        raise RuntimeError
-
-    def get_byte(self, key):
-        val = int(self.real[key]*255.99)
-        assert val < 256 and val >= 0
-        return val
-
-    def set_byte(self, key, val):
-        if val >= 256 or val < 0:
-            raise ValueError
-        if key not in self.real:
-            raise KeyError
-        self.real[key] = val/255.99
-
-    def set_bytes(self, **kwargs):
-        for key, value in kwargs.items():
-            if type(value) is bool:
-                raise ValueError
-            else:
-                self.set_byte(key, value)
-
-    def get_color(self):
-        return (
-            self.get_byte("red"),
-            self.get_byte("green"),
-            self.get_byte("blue"),
-            255
-        )
-
-    def randomize(self):
-        for key in self.binary:
-            self.binary[key] = (random.random() >= 0.5)
-        for key in self.real:
-            self.real[key] = random.random()
-
-    def get_mapped_real(self, key, start, stop):
-        return ((stop-start)*self.real[key] + start)
-
 # TODO: Separate out a cluster object class for multiple rect objects like this
 class Robot(PhysicsRectangle):
     def __init__(self, world, *args, stroke=(0,0,0,255), fill=(128,128,128,255),dna=None, **kwargs):
         if not dna:
-            self.dna = RobotDNA()
+            self.dna = DNA()
         else:
             self.dna = dna
         self.world = world
